add listing issue finding to helper fn
